Refactor OVS -> JIT
