+++ .github/workflows/validateJITMaster.yml
@@ -1,4 +1,4 @@
-name: Run testsuite on OVS master
+name: Run testsuite on JIT master
 
 on: [push]
 
@@ -40,13 +40,13 @@ jobs:
     - name: Load DCSA test data
       run: cat DCSA-OVS/DCSA-Information-Model/datamodel/testdata.d/*.sql | docker exec -i -u postgres dcsa_db psql -d dcsa_openapi -f -
 
-    - name: Run the API validator on OVSV2
-      run: cd DCSA-API-Validator && mvn -Dtest.suite=OVSV2.xml test
+    - name: Run the API validator on JITV1
+      run: cd DCSA-API-Validator && mvn -Dtest.suite=JITV1.xml test
       env: 
         client_secret: ${{ secrets.AUTH0_DCSA_CLIENT_SECRET }}
         client_id: 1qEYXCbxWvjeYS8VVm1BfnfZrP3TY9n8
         audience: localhost
         OAuthTokenUri: https://dcsa.eu.auth0.com/oauth/token
-        API_ROOT_URI: http://localhost:9090/v2
+        API_ROOT_URI: http://localhost:9090/v1
         CALLBACK_URI: http://172.17.0.1:4567
         DCSA_API_VALIDATOR_MAY_USE_POST_EVENTS_ENDPOINT: true
