+++ JITV1.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
 <suite name="Suite">
-    <test thread-count="5" name="OVSV2">
+    <test thread-count="5" name="JITV1">
         <packages>
-            <package name="org.dcsa.api_validator.ovs.v2"/>
+            <package name="org.dcsa.api_validator.jit.v1"/>
             <package name="org.dcsa.api_validator.conf"/>
         </packages>
     </test> <!-- Test -->
