+++ src/test/resources/jit/v1/EventSubscription/InvalidEventTypeEventSubscriptionSample.json
@@ -1,8 +1,7 @@
 {
-  "callbackUrl": "http://127.0.0.1:9092/v2/notification-endpoints/receive/37eb1b03-db67-4b57-abb8-d38c3faed703",
+  "callbackUrl": "http://127.0.0.1:9092/v1/notification-endpoints/receive/37eb1b03-db67-4b57-abb8-d38c3faed703",
   "eventType": [
-    "TRANSPORT",
-    "OPERATIONS"
+    "SHIPMENT"
   ],
   "transportEventTypeCode": [
     "ARRI",
