+++ src/test/resources/jit/v1/EventsSchema.json
